fix: broken type checks as of Live 12.1 (#7)
Live 12.1 switched to python 3.11, which is currently impractical to
decompile. See
https://github.com/gluon/AbletonLive12_MIDIRemoteScripts/issues/2 for
more discussion.

`pycdc` does decompile some basic structural elements of the Live
libraries, but outputs mostly empty function and class definitions.
https://pylingual.io/ seems to work better but isn't currently possible
to integrate in an automated way.

This PR adds https://github.com/gluon/AbletonLive12_MIDIRemoteScripts
(currently not updated for 12.1 but seems to work fine for our use
cases) as a submodule and sets it up for use with the typechecker.